refactor v2 parser pipe tests
